+++ cuda/include/particles.h
@@ -2,8 +2,8 @@
 #define DSMC_PARTICLES_H
 
 typedef struct {
-    double *x, *y, *z;
-    double *vx, *vy, *vz;
+    float *x, *y, *z;
+    float *vx, *vy, *vz;
 } Particles;
 
 #endif // DSMC_PARTICLES_H
